+++ content-types/content-type-primitives/src/index.ts
@@ -1,72 +1,50 @@
-import type { content } from "@xmtp/proto";
+import type { ContentTypeId, EncodedContent } from "@xmtp/node-bindings";
 
-export class ContentTypeId {
-  authorityId: string;
-
-  typeId: string;
-
-  versionMajor: number;
-
-  versionMinor: number;
-
-  constructor(obj: content.ContentTypeId) {
-    this.authorityId = obj.authorityId;
-    this.typeId = obj.typeId;
-    this.versionMajor = obj.versionMajor;
-    this.versionMinor = obj.versionMinor;
-  }
-
-  toString(): string {
-    return `${this.authorityId}/${this.typeId}:${this.versionMajor}.${this.versionMinor}`;
-  }
-
-  static fromString(contentTypeString: string): ContentTypeId {
-    const [idString, versionString] = contentTypeString.split(":");
-    const [authorityId, typeId] = idString.split("/");
-    const [major, minor] = versionString.split(".");
-    return new ContentTypeId({
-      authorityId,
-      typeId,
-      versionMajor: Number(major),
-      versionMinor: Number(minor),
-    });
-  }
-
-  sameAs(id: ContentTypeId): boolean {
-    return this.authorityId === id.authorityId && this.typeId === id.typeId;
-  }
-}
-
-export type EncodedContent<Parameters = Record<string, string>> = {
-  type: ContentTypeId;
-  parameters: Parameters;
-  fallback?: string;
-  compression?: number;
-  content: Uint8Array;
-};
-
-export type ContentCodec<
-  ContentType = unknown,
-  Parameters = Record<string, string>,
-> = {
+export type ContentCodec<ContentType = unknown> = {
   contentType: ContentTypeId;
-  encode(
-    content: ContentType,
-    registry: CodecRegistry,
-  ): EncodedContent<Parameters>;
-  decode(
-    content: EncodedContent<Parameters>,
-    registry: CodecRegistry,
-  ): ContentType;
+  encode(content: ContentType): EncodedContent;
+  decode(content: EncodedContent): ContentType;
   fallback(content: ContentType): string | undefined;
   shouldPush: (content: ContentType) => boolean;
 };
 
 /**
- * An interface implemented for accessing codecs by content type.
+ * Compares two content type IDs for equality
+ *
+ * @param a - First content type ID
+ * @param b - Second content type ID
+ * @returns True if the content type IDs are equal
  */
-export interface CodecRegistry<T = unknown> {
-  codecFor(contentType: ContentTypeId): ContentCodec<T> | undefined;
-}
+export const contentTypesAreEqual = (a: ContentTypeId, b: ContentTypeId) =>
+  a.authorityId === b.authorityId && a.typeId === b.typeId;
 
-export type CodecMap<T = unknown> = Map<string, ContentCodec<T>>;
+/**
+ * Converts a content type ID to a string
+ *
+ * @param contentType - Content type ID
+ * @returns String representation of the content type ID
+ */
+export const contentTypeToString = (contentType: ContentTypeId) =>
+  `${contentType.authorityId}/${contentType.typeId}:${contentType.versionMajor}.${contentType.versionMinor}`;
+
+/**
+ * Converts a string to a content type ID
+ *
+ * @param contentTypeString - String representation of the content type ID
+ * @returns Content type ID
+ */
+export const contentTypeFromString = (
+  contentTypeString: string,
+): ContentTypeId => {
+  const [idString, versionString] = contentTypeString.split(":");
+  const [authorityId, typeId] = idString.split("/");
+  const [major, minor] = versionString.split(".");
+  return {
+    authorityId,
+    typeId,
+    versionMajor: Number(major),
+    versionMinor: Number(minor),
+  };
+};
+
+export type { ContentTypeId, EncodedContent };